// src/lib/slider/enums/directions.ts
/**
 * Axis of the arrangement of a slider.
 * @link https://github.com/DjonnyX/centrifugal/blob/main/src/lib/slider/enums/directions.ts
 * @author Evgenii Alexandrovich Grebennikov
 * @email djonnyx@gmail.com
 */
export enum Directions {
    /**
     * Horizontal axis.
     */
    HORIZONTAL = 'horizontal',
    /**
     * Vertical axis.
     */
    VERTICAL = 'vertical',
}